requirements.txt
fastapi
uvicorn[standard]
google-cloud-speech
moviepy
pydub
boto3
Pillow
requests
openai
